+++ thumbs/thumbs.py
@@ -21,8 +21,8 @@ class ThumbsBlock(InputBlock):
 
     """
 
-    upvotes = Integer(help="Number of up votes", default=0, scope=Scope.content)
-    downvotes = Integer(help="Number of down votes", default=0, scope=Scope.content)
+    upvotes = Integer(help="Number of up votes", default=0, scope=Scope.user_state_summary)
+    downvotes = Integer(help="Number of down votes", default=0, scope=Scope.user_state_summary)
     voted = Boolean(help="Has this student voted?", default=False, scope=Scope.user_state)
 
     def student_view(self, context):  # pylint: disable=W0613